Cleaned up includes. Removed duplicate code in LoadHits(). Fixed bug where two PFPs in different cryostats are assigned the same color/grouping. Moved global vars to plot_tools.h
